修复bug

+++ stock/shares/model/shares_block_gns.py
@@ -9,11 +9,11 @@ from datetime import datetime
 class SharesBlockGns(models.Model):
     code = models.ForeignKey(SharesName, name='code', on_delete=models.CASCADE)
     name = models.CharField(max_length=30)
-    p_min = models.IntegerField(help_text="最小")
-    p_max = models.IntegerField(help_text="最大")
-    p_start = models.IntegerField(help_text="开始")
-    p_end = models.IntegerField(help_text="结束")
-    p_zhang_die_fu = models.IntegerField(help_text="涨跌幅")
+    p_min = models.FloatField(help_text="最小")
+    p_max = models.FloatField(help_text="最大")
+    p_start = models.FloatField(help_text="开始")
+    p_end = models.FloatField(help_text="结束")
+    p_zhang_die_fu = models.FloatField(help_text="涨跌幅")
     date_as = models.DateField(help_text="创建时间")
 
     class Meta:
